+++ dash/examples/ex.04.memalloc/main.cpp
@@ -9,31 +9,31 @@ using namespace std;
 int main(int argc, char* argv[])
 {
   dash::init(&argc, &argv);
-  
+
   auto myid = dash::myid();
   auto size = dash::size();
 
   dash::Array< dash::GlobPtr<int> > arr(size);
 
   arr[myid] = dash::memalloc<int>(SIZE);
-  
+
   for(int i=0; i<SIZE; i++ ) {
     dash::GlobPtr<int> ptr = arr[myid];
     ptr[i]=myid;
   }
-  
+
   dash::barrier();
 
   cout<<myid<<": ";
   for(int i=0; i<SIZE; i++ ) {
     dash::GlobPtr<int> ptr = arr[(myid+1)%size];
-    cout<<ptr[i]<<" ";
+    cout<<(int)ptr[i]<<" ";
   }
   cout<<endl;
 
   dash::barrier();
 
   //dash::memfree(arr[myid]);
-  
+
   dash::finalize();
 }
